+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-26T14:56:52.220Z",
+    "2026-02-27T09:23:51.039Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {