+++ packages/orq-rc/RELEASES.md
@@ -7349,3 +7349,13 @@ Based on:
 - [typescript v4.5.0-rc.14] packages/orq-rc
 ### Releases
 - [NPM v4.5.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.14 - packages/orq-rc
+
+## 2026-02-27 09:26:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.731.2 (2.845.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.5.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.5.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.5.0-rc.15 - packages/orq-rc